further organize functions folder by adding helpers folders, rename functions for readability
